Moving to a new repository.
The old one was over 1.2 GB, while the new one is under 5 MB!

The reason the old one was so big was because of my lack of git knowledge.
Back in 2014 I committed all of the music that I wanted to use in my game,
thinking that git was sort of like Dropbox or Google Drive. I've tried for years
to remove them from my git history, and I've finally given up.

I'm going to keep the old repo around for documentation, and because I'm very
nostalgic about it, but new commits will probably only go to this new repo.