Programmed StatusPotions. Reworked how items are used.
Fixed the Unit.Speed stat not doing anything (whoops...)
Cleaned up a lot of code.

There is currently I bug that I need to fix ASAP: if you use an item in battle, it doesn't disappear until
the item is actually used. This means that multiple party members can use the same item on one turn,
effectively duplicating it.