Added some new dialogue functionality.
You can now have NPCs say the player's name, using the new TEXT_TO_REPLACE property.
If you include this text in a string, it will auto-replace with your character's name when the NPC
talks to you.

Also added a line of whitespace between dialogues.
Also fixed up Philliard's dialogue, and added some new dialogue for him.

Overall I'm really happy with how I'm doing the dialogue system. It's a lot cleaner and simpler than
the old engine did it, and I think it's going to be much easier to work with.