Made many improvements to the engine.
Fixed some memory leaks with sounds and music.

TempStats is now a StatMatrix instead of a Dictionary<string, int>. This means that you can access the values
by typing e.g. unit.TempStats.Attack instead of unit.TempStats["attack"], which is identical to how the true stats are accessed. This also prevents typo errors such as attempting to access unit.TempStats["attk"] which would only throw an error on runtime.

Moved many variables to GameLoopManager and MusicPlayer.

Reworked the CalculateDamage() method, it's now been split into three different methods that are more specialized:
CalculateAttackDamage() is used for attacks that are amplfied by Weapon Power, such as Standard Attacks or some abilities.
CalculateSpellDamage() is used for attacks that are amplified by Charisma, such as Spells and maybe some abilities.
CalculateRawDamage() is used for attacks that are not amplified by anything, such as throwing weapons or fixed-damage attacks.