Tons of code cleanup. Also preparing to add bosses to the game.
First step of the process was removing the Melee/Ranged/MagicMonster classes and implementing their
functionality in some other way, since it is impossible to have bosses inherit from both a Boss class and a
Melee/Ranged/MagicMonster class at the same time. The new method is arguably cleaner anyway since
all the AI is in the same place. I am planning on making additional changes to the way the AI works.